Refactored to utilize Observer-Subject Pattern
Feel free to paraphrase this in your report if we end up using this version:
Two interfaces have been created:
- Subject which contains notify(event), which passes an event enum.
- Observer, which contains update(event), which performs an action in response to the event.

GUI, ActionPanel, and GameSetupFrame all implement Subject and essentially act as controllers. 
MurderMadness implements Observer.
Each Subject class contains a 'list' of Observers (only one in this case because they only need to communicate with one other class).
When a Subject needs to send an instruction to an Observer, it uses notify(event) which iterates through all the observers in the list and tells them to update(event).
The Event enum determines what the Observer will do, so the Observers that respond to the particular enum will perform that action.

The main purpose of this pattern is to achieve decoupling, such that changes in one Subject do not necessarily affect the Observers, and vice versa.